% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{calculate_sizes}
\alias{calculate_sizes}
\title{Calculate Sizes}
\usage{
calculate_sizes(df)
}
\arguments{
\item{df}{a data frame of libraries}
}
\value{
a data frame with a lib_size column
}
\description{
Calculate Sizes
}
