% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{calculate_sizes}
\alias{calculate_sizes}
\title{calculate sizes}
\usage{
calculate_sizes(df)
}
\arguments{
\item{df}{a dataframe of libraries}
}
\value{
a \code{data.frame} with a lib_size column
}
\description{
calculate sizes
}
