% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib}
\alias{lib}
\title{Generate a data frame of installed packages}
\usage{
lib()
}
\value{
a data.frame of all packages installed on your system
}
\description{
Generate a data frame of installed packages
}
