% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib}
\alias{lib}
\title{Generate a dataframe of installed packages}
\usage{
lib()

lib()
}
\value{
dataframe of all packages installed on a system
@export

a data.frame of all packages installed on your system
}
\description{
Generate a dataframe of installed packages

Generate a data frame of installed packages
}
