% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide Number of R packages by library}
\usage{
lib_summary()
}
\value{
a data.frame of R packages by library
}
\description{
Provide Number of R packages by library
}
\examples{
lib_summary()
}
