% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R library summary}
\usage{
lib_summary()
}
\value{
a data.frame containing the count of pacakges in each of
your libraries
}
\description{
Provides a brief summary of package libraries
}
\examples{
lib_summary()
}
