% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(size = FALSE)
}
\arguments{
\item{size}{A logical value to indicate whether to calcluate library sizes. Default \code{FALSE}}
}
\value{
A \code{data.frame} containing the count of packages in each
of the user's libraries. If \code{size = TRUE}, a column of library sizes will be added.
}
\description{
Provide a brief summary of the package libraries on
your machine
}
\examples{
lib_summary()
lib_summary(size = TRUE)
}
