% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary()
}
\value{
A \code{data.frame} of library paths and total count of packages contained within.
}
\description{
Provides utility functions for inspecting and examining your local R package installations.
}
\examples{
lib_summary()
}
