% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary
BLAH BLAH BLAH}
\usage{
lib_summary()
}
\value{
a \code{data.frame}  blah
}
\description{
R Library Summary
BLAH BLAH BLAH
}
\examples{
lib_summary()
}
