% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary
Provides a brief summary of the package libraries
on your machine
@param sizes A logical value indicating whether to calculate sizes
Default \code{FALSE}}
\usage{
lib_summary(sizes = FALSE)
}
\value{
a \code{data.frame}  containing the count of packages in each of the
user's libraries.  If \code{sizes=TRUE}, a column of
library sizes will be added.
}
\description{
R Library Summary
Provides a brief summary of the package libraries
on your machine
@param sizes A logical value indicating whether to calculate sizes
Default \code{FALSE}
}
\examples{
lib_summary()
lib_summary(sizes=TRUE)
}
