% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide number of R Packages by library}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{Should sizes of libraries be calculated. Default \code{FALSE}}
}
\value{
A data.frame containing the count of packages by library in each of
the users libraries
}
\description{
Provide the number of R packages by library in a data.frame
}
\examples{
lib_summary()
}
