% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Enumerate installed R packages}
\usage{
lib_summary()
}
\value{
A data.frame containing the count of packages
in each of your libraries.
}
\description{
Searches for and enumerates R packages installed in various locations on your local machine.
It takes no arguments.
}
\examples{
\dontrun{
lib_summary()
}
}
