% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide Number of R packages by Library}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{should sizes of libraries be calculated. Default 'FALSE'}
}
\value{
a data.frame containing the count of packages in each of the user's libraries
}
\description{
Provide the number of R packages by library in a data frame
}
\examples{
lib_summary()
}
