% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R library summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{A logical value indicating weather to calculate sizes
default \code{FALSE}}
}
\value{
A \code{data.frame} containing the count of packages in each
of the user's libraries
}
\description{
Provides a brief summary of package libraries on
your machine
}
\examples{
lib_summary()
lib_summary(sizes = TRUE)
}
