% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Summarizes R Libraries}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{logical indicating whether or not to calculate sizes of libraries. Default \code{FALSE}.}
}
\value{
A \code{data.frame} containing the count of packages in each of the user's
libraries
}
\description{
Summarizes R libraries on the local machine, including number of
packages installed.
}
\examples{
lib_summary()
}
