% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Summary of libraries}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{TRUE / FALSE to assess the size of libraries}
}
\value{
A dataframe containing count of packages with libraries
}
\description{
Provides a list of number of packages per library
location on your machine
}
\examples{
\dontrun{
lib_summary()
lib_summary(sizes = TRUE)
}
}
