% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Summary of Libraries}
\usage{
lib_summary()
}
\value{
A dataframe containing count of packages with libraries
}
\description{
Provides a list of number of packages per library
location on your machine
}
\examples{
\dontrun{
lib_summary()
}
}
