% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Library Summary
Provides a brief summary of the package libraries on your computer}
\usage{
lib_summary()
}
\value{
a data.frame containing the count of packages
}
\description{
Library Summary
Provides a brief summary of the package libraries on your computer
}
\examples{
lib_summary()
}
