% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{Should sizes of libraries be calculated. Default 'FALSE'.}
}
\value{
a data.frame of R packages by library
}
\description{
Provides a brief summary of the package libraries on your machine
}
\examples{
lib_summary()
}
