% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R libary summary}
\usage{
lib_summary()
}
\value{
A data frame containing the paths and package counts for each library
}
\description{
Provides a summary of libraries and numbers of packages installed
}
\examples{

lib_summary()

}
