% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{logical indicating whether or not to calculate library sizes. Default \code{FALSE}.
Provide a brief summary of your the package libraries on your machine}
}
\value{
A \code{data.frame} containing the count of packages in each of the
user's libraries
}
\description{
R Library Summary
}
\examples{
lib_summary()
}
