% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = F)
}
\arguments{
\item{sizes}{logical indicating whether or not to calculate library sizes
Default \code{FALSE}}
}
\value{
A \code{data.frame} containing count of packages per library and location
of user libraries
}
\description{
Provide a brief summary of package libraries on machine
}
\examples{
lib_summary()
}
