% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R library summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{Should sizes of libraries be calculated. Default \code{FALSE}}
}
\value{
A data frame containing the count of packages in each of the user's library
}
\description{
R library summary
}
