% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide number of R packages by Library}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{Should sizes of libraries be calculated. Default \code{FALSE} .}
}
\value{
a data.frame of the number of packages in Program Files and Users library
}
\description{
Provide number of R packages by Library
}
\examples{
lib_summary()
}
