% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = FALSE)
}
\value{
A \code{data.frame} containing a count of the packages in each of the user's
libraries.
}
\description{
Provides a brief summary of the packages installed on your machine.
}
\examples{
lib_summary()
lib_summary(sizes = TRUE)
}
