% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{A function title}
\usage{
lib_summary(sizes = F)
}
\arguments{
\item{sizes}{LOGICAL; if TRUE (default), adds a column of the total bit size of libraries}
}
\value{
a data.frame of packages called
}
\description{
A function description
}
\examples{
lib_summary()
}
