% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide number of R packages by library}
\usage{
lib_summary(by = .data$LibPath, sizes = FALSE)
}
\arguments{
\item{by}{tidy-select columns to group by. Defaults to \code{LibPath}}

\item{sizes}{logical should file sizes be calculated? Defaults
to \code{FALSE}}
}
\value{
a data.frame of R packages by library
}
\description{
Provide the number of R packages by library in a data frame
}
\examples{
lib_summary()
}
