% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{A logical value indicated whether to caclulate sizes. Default 'FALSE'.}
}
\value{
A 'data.frame' containing the count of packages. If 'sizes = TRUE', a column of library sizes will be added.
}
\description{
Provides a brief summary of package libraries.
}
\examples{
lib_summary()
lib_summary(sizes = TRUE)
}
