% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary()
}
\value{
A \code{data.frame} containing a count of packages in eachof the user's
libraries.
}
\description{
Provides a brief summary of the pakcage libraries
on your machine
}
\examples{
lib_summary()
}
