% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Summarise Installed Package Libraries}
\usage{
lib_summary()
}
\value{
A data frame with columns \code{Library} and \code{n_packages}.
}
\description{
Count installed packages by library path.
}
\examples{
lib_summary()
}
