% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide Number of R Packages by Library}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{Should sizes of libraries be calculated. Default 'FALSE'}
}
\value{
A data.frame containing the count of packages in each of the user's R libraries
}
\description{
Provide the number of R package by library in
a data.frame
}
\examples{
lib_summary
}
