% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide Number of R Packages by Library
R Library Summary}
\usage{
lib_summary(by = .data$LibPath, sizes = FALSE)
}
\arguments{
\item{by}{The grouping variable, grouped by LibPath, i.e. by library path}

\item{sizes}{Should sizes of libraries be calculated. Default \code{FALSE}.}
}
\value{
A data.frame containing the count of packages in each of the user's
libraries
}
\description{
Provides a brief summary of the package libraries on your machine
}
\examples{
lib_summary()
}
