% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Library Summary
Provide a brief summary of the package
libraries on your machine}
\usage{
lib_summary()
}
\value{
A\verb{data frame} containing the count of packages in each of the user's libs
}
\description{
Library Summary
Provide a brief summary of the package
libraries on your machine
}
\examples{
lib_summary()
}
