% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{Should the sizes of the libraries be calculated?}
}
\value{
A data.frame containing the count of packages in each of the
user's libraries. A \code{lib_size} column is included if \code{sizes = TRUE}.

A \code{data.frame} containing the count of packages in each of the
user's libraries
}
\description{
Provides a brief summary of the package libraries on your machine
}
\examples{
lib_summary()
}
