% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Libraries Summary}
\usage{
lib_summary()
}
\value{
A data.frame containing a count of packages in
each of your libraries
}
\description{
Provides a brief summary of the locally-installed
R libraries
}
\examples{
lib_summary()
}
