% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{provide number of r packages by lirbary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{should sizes of libraries be calculated, default is false}
}
\value{
a data frame of R
}
\description{
provide number of r packages by lirbary
}
\examples{
lib_summary()
}
