% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{provide number of r packages by lirbary}
\usage{
lib_summary()
}
\value{
a data frame of R
}
\description{
provide number of r packages by lirbary
}
\examples{
lib_summary()
}
