% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{Provide the number of R Packages}
\usage{
lib_summary()
}
\value{
a data.frame  of R packages
}
\description{
Provide the number of R Packages
}
\examples{
lib_summary()
}
