% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{logical indicating whether or not to calculate library
sizes. Default \code{FALSE}}
}
\value{
A \code{data.frame} containing the counts of packages in each of the
user's libraries
}
\description{
Provide a brief summary of the package libraries on your machine.
}
\examples{
lib_summary()
}
