% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary()
}
\value{
A \code{data.frame} containing the counts of pacakgages in each of the
user's libraries
}
\description{
Provide a brief summary of the package libararies on your machine.
}
\examples{
lib_summary()
}
