% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Exploring how to build an R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/a-potapova/libminer}
  \item Report bugs at \url{https://github.com/a-potapova/libminer/issues}
}

}
\author{
\strong{Maintainer}: Anna Potapova \email{annaa.potapova@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
