% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: What the Package Does (One Line, Title Case)}
\description{
Provides files for learning about your own libraries, and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aclink88/libminer}
  \item \url{https://aclink88.github.io/libminer/}
  \item Report bugs at \url{https://github.com/aclink88/libminer/issues}
}

}
\author{
\strong{Maintainer}: Reginald Buckingshire \email{aclink85@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
