% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adsmithca/libminer}
  \item \url{https://adsmithca.github.io/libminer/}
  \item Report bugs at \url{https://github.com/adsmithca/libminer/issues}
}

}
\author{
\strong{Maintainer}: Andrew Smith \email{andrew.d.smith.ca@gmail.com}

}
\keyword{internal}
