% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R packages}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agneke/libminer}
  \item Report bugs at \url{https://github.com/agneke/libminer/issues}
}

}
\author{
\strong{Maintainer}: Agne Velthut-Meikas \email{agne.velthut@taltech.ee}

Authors:
\itemize{
  \item Agne Velthut-Meikas \email{agne.velthut@taltech.ee}
}

}
\keyword{internal}
