% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexnhdg/libminer}
  \item \url{https://alexnhdg.github.io/libminer/}
  \item Report bugs at \url{https://github.com/alexnhdg/libminer/issues}
}

}
\author{
\strong{Maintainer}: Alexander de Groot \email{alexnhdg@gmail.com}

}
\keyword{internal}
