% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore local R libraries}
\description{
Provides functions for learning about libraries on your local drive.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amason30/libminer}
  \item \url{https://amason30.github.io/libminer/}
  \item Report bugs at \url{https://github.com/amason30/libminer/issues}
}

}
\author{
\strong{Maintainer}: Allison Mason \email{masonar@umich.edu} (\href{https://orcid.org/0000-0003-1339-1592}{ORCID})

}
\keyword{internal}
