% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides fucntions for learning about your R libraries and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amessay2/libminer}
  \item Report bugs at \url{https://github.com/amessay2/libminer/issues}
}

}
\author{
\strong{Maintainer}: Endegena Desta \email{andymessay@gmail.com}

}
\keyword{internal}
