% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anacarlac/libminer}
  \item Report bugs at \url{https://github.com/anacarlac/libminer/issues}
}

}
\author{
\strong{Maintainer}: Ana Crispim \email{anacarlacrispim@gmail.com} (\href{https://orcid.org/0000-0003-3434-5667}{ORCID})

}
\keyword{internal}
