% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Summarize Your R Libraries}
\description{
This package summarizes your R libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrea-raisman/libminer}
  \item Report bugs at \url{https://github.com/andrea-raisman/libminer/issues}
}

}
\author{
\strong{Maintainer}: Andrea Villanueva Raisman \email{andrea.villanueva@gmail.com}

}
\keyword{internal}
