% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ann-ced/libminer}
  \item Report bugs at \url{https://github.com/ann-ced/libminer/issues}
}

}
\author{
\strong{Maintainer}: Anna Cederberg \email{anna.cederberg@gmail.com}

Authors:
\itemize{
  \item Anna Cederberg \email{anna.cederberg@gmail.com}
}

}
\keyword{internal}
