% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Summarize R packages on your computer}
\description{
Summarize R packages on your computer (toy package).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ateucher/libminer}
  \item \url{http://andyteucher.ca/libminer/}
  \item Report bugs at \url{https://github.com/ateucher/libminer/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

}
\keyword{internal}
