% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Examine your R libraries!}
\description{
Provides utility functions for inspecting and examining your local R package installations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/awesomecosmos/libminer}
  \item Report bugs at \url{https://github.com/awesomecosmos/libminer/issues}
}

}
\author{
\strong{Maintainer}: Aayushi Verma \email{theawesomecosmos@gmail.com} (\href{https://orcid.org/0000-0003-2396-4569}{ORCID})

}
\keyword{internal}
