% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides funtions for learning about your R libraires and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bauersam/libminer}
  \item Report bugs at \url{https://github.com/bauersam/libminer/issues}
}

}
\author{
\strong{Maintainer}: Samantha Bauer \email{sbauer@wayne.edu} (\href{https://orcid.org/0000-0002-2529-2930}{ORCID})

}
\keyword{internal}
