% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about R libraries and packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/betsymccoach/libminer}
  \item \url{https://betsymccoach.github.io/libminer/}
  \item Report bugs at \url{https://github.com/betsymccoach/libminer/issues}
}

}
\author{
\strong{Maintainer}: Betsy McCoach \email{betsy@uconn.edu} (\href{https://orcid.org/0000-0001-9063-6835}{ORCID})

}
\keyword{internal}
