% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Fundamentals of Package Development Example}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/deeprich/libminer}
  \item Report bugs at \url{https://github.com/deeprich/libminer/issues}
}

}
\author{
\strong{Maintainer}: Richard Johnson \email{public@richardjohnson.iocheck} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
