% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Summarise R packages on your computer.}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/catarinawor/libminer}
  \item Report bugs at \url{https://github.com/catarinawor/libminer/issues}
}

}
\author{
\strong{Maintainer}: Catarina Wor \email{catarina.wor@dfo-mpo.gc.ca}

}
\keyword{internal}
