% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Package Library}
\description{
Enumerate the number of downloaded R packages on your local machine; searches multiple directories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chrispmad/libminer}
  \item Report bugs at \url{https://github.com/chrispmad/libminer/issues}
}

}
\author{
\strong{Maintainer}: Chris Madsen \email{chris.madsen@gov.bc.ca}

}
\keyword{internal}
