% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Librarues}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\author{
\strong{Maintainer}: Lee Zeman \email{lzeman@uvic.ca}

}
\keyword{internal}
