% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
This package provides functions that allow users to explore their R libraries and the pacakges therein.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coughlincarter/libminer}
  \item \url{https://coughlincarter.github.io/libminer/}
  \item Report bugs at \url{https://github.com/coughlincarter/libminer/issues}
}

}
\author{
\strong{Maintainer}: Carter Coughlin \email{cocoughlin@me.com}

}
\keyword{internal}
