% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Writes functions for learning about your R libraries, and the packags they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/darbycamp/libminer}
  \item Report bugs at \url{https://github.com/darbycamp/libminer/issues}
}

}
\author{
\strong{Maintainer}: Darby Comstock \email{darbycampbell14@gmail.com}

}
\keyword{internal}
