% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Expolre R libraries}
\description{
This package contains fun functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edgar-vestcor/libminer}
  \item \url{https://edgar-vestcor.github.io/libminer/}
  \item Report bugs at \url{https://github.com/edgar-vestcor/libminer/issues}
}

}
\author{
\strong{Maintainer}: Jivitesh Gudekar \email{jivitesh.gudekar@vestcor.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
