% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiheather/libminer}
  \item Report bugs at \url{https://github.com/epiheather/libminer/issues}
}

}
\author{
\strong{Maintainer}: Heather Lee \email{HLEE73@mgh.harvard.edu} (\href{https://orcid.org/0000-0001-7517-6594}{ORCID})

}
\keyword{internal}
