% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Utility for Inspecting Libraries}
\description{
Provides functions to learn about R libraries in your system, including the packages that are installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ericgyoung/libminer}
  \item Report bugs at \url{https://github.com/ericgyoung/libminer/issues}
}

}
\author{
\strong{Maintainer}: Eric Young \email{ericgalenyoung@gmail.com}

}
\keyword{internal}
