% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R libraries}
\description{
Provides helper functions to explore and examine your package library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erwinlares/libminer}
  \item Report bugs at \url{https://github.com/erwinlares/libminer/issues}
}

}
\author{
\strong{Maintainer}: Erwin Lares \email{erwin.lares@wisc.edu} (\href{https://orcid.org/0000-0002-3284-828X}{ORCID})

}
\keyword{internal}
