% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Expelor Your R libraries}
\description{
Provides functions for learning about your R libraries.
}
\author{
\strong{Maintainer}: Fatemeh Hatefi \email{f.hatefi@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
