% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries and the packages that they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franzenr/libminer}
  \item Report bugs at \url{https://github.com/franzenr/libminer/issues}
}

}
\author{
\strong{Maintainer}: Rose Franzen \email{franzenr@chop.edu} (\href{https://orcid.org/0000-0001-5750-6473}{ORCID})

}
\keyword{internal}
