% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FulyaGokalp2/libminer}
  \item Report bugs at \url{https://github.com/FulyaGokalp2/libminer/issues}
}

}
\author{
\strong{Maintainer}: Fulya Gokalp Yavuz \email{fulyagokalp@hotmail.com} (\href{https://orcid.org/0000-0002-7750-9767}{ORCID})

}
\keyword{internal}
