% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Check Libraries On Computer}
\description{
Help explore the libraries on your machine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gcperk/libminer}
  \item \url{https://gcperk.github.io/libminer/}
  \item Report bugs at \url{https://github.com/gcperk/libminer/issues}
}

}
\author{
\strong{Maintainer}: Genevieve Perkins \email{genevieve.c.perkins@gmail.com}

}
\keyword{internal}
