% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R libraries}
\description{
Provides functions for learning about your R libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iammrtza/libminer}
  \item Report bugs at \url{https://github.com/iammrtza/libminer/issues}
}

}
\author{
\strong{Maintainer}: Morteza Aslanzadeh \email{morteza.aslanzadeh@ki.se}

Authors:
\itemize{
  \item Morteza Aslanzadeh \email{morteza.aslanzadeh@ki.se}
}

}
\keyword{internal}
