% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R library and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iaradsouza1/libminer}
  \item Report bugs at \url{https://github.com/iaradsouza1/libminer/issues}
}

}
\author{
\strong{Maintainer}: Iara Souza \email{iaradsouza1@gmail.com} (\href{https://orcid.org/0000-0002-2550-6150}{ORCID})

}
\keyword{internal}
