% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Explores R libraries in your R setup.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/InfiniteCuriosity/libminer}
  \item Report bugs at \url{https://github.com/InfiniteCuriosity/libminer/issues}
}

}
\author{
\strong{Maintainer}: Russ Conte \email{russconte@mac.com}

}
\keyword{internal}
