% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R packages - for RaukR course 2025}
\description{
Function to find path to your R libraries and how many packages you have installed. Original package created by ateucher.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbjurling/libminer}
  \item Report bugs at \url{https://github.com/jbjurling/libminer/issues}
}

}
\author{
\strong{Maintainer}: Josefin Bjurling \email{bjurlingjosefin@gmail.com} (\href{https://orcid.org/0009-0006-7675-429X}{ORCID})

}
\keyword{internal}
