% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R Packages}
\description{
Provides a framework for describing packages installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JMarkel23/libminer}
  \item Report bugs at \url{https://github.com/JMarkel23/libminer/issues}
}

}
\author{
\strong{Maintainer}: Justin Markel \email{justinm@ping.com}

}
\keyword{internal}
