% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Mining R libraries}
\description{
Mining R libraries.
}
\author{
\strong{Maintainer}: Jon Duan \email{jon.duan@gov.bc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
