% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore R Libraries}
\description{
Provides functions for learning about R libraries and packages installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/k-maciejewski/libminer}
  \item Report bugs at \url{https://github.com/k-maciejewski/libminer/issues}
}

}
\author{
\strong{Maintainer}: Kaitlin Maciejewski \email{katemac68@gmail.com}

}
\keyword{internal}
