% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R Libraries}
\description{
summarize libraries by packages. In a course by Andy Teucher.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kateaurora/libminer}
  \item \url{https://kateaurora.github.io/libminer/}
  \item Report bugs at \url{https://github.com/kateaurora/libminer/issues}
}

}
\author{
\strong{Maintainer}: Kate McGivney \email{katmcgivney@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
