% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaylgill/libminer}
  \item \url{https://kaylgill.github.io/libminer/}
  \item Report bugs at \url{https://github.com/kaylgill/libminer/issues}
}

}
\author{
\strong{Maintainer}: Kayleigh Gillespie \email{Kayleigh.Gillespie@dfo-mpo.gc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
