% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the 
    packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinyang25/libminer}
  \item Report bugs at \url{https://github.com/kevinyang25/libminer/issues}
}

}
\author{
\strong{Maintainer}: Kevin Yang \email{kjy110498@gmail.com}

}
\keyword{internal}
