% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides funcions for learning abou your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kim-s-h/libminer}
  \item Report bugs at \url{https://github.com/Kim-s-h/libminer/issues}
}

}
\author{
\strong{Maintainer}: Son Kim \email{sseohyun@gmail.com} (\href{https://orcid.org/0000-0002-2119-6044}{ORCID})

}
\keyword{internal}
