% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Returns the Number of Packages in Each of Your Libraries}
\description{
Returns the Number of Packages in Each of Your Libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lgoris/libminer}
  \item Report bugs at \url{https://github.com/lgoris/libminer/issues}
}

}
\author{
\strong{Maintainer}: Lisse Goris \email{lsse.goris@su.se}

}
\keyword{internal}
