% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lind-i/libminer}
  \item Report bugs at \url{https://github.com/lind-i/libminer/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
