% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions fo learning about your R libraries and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmaowisc/libminer}
  \item \url{https://lmaowisc.github.io/libminer/}
  \item Report bugs at \url{https://github.com/lmaowisc/libminer/issues}
}

}
\author{
\strong{Maintainer}: Lu Mao \email{lmao@biostat.wisc.edu} (\href{https://orcid.org/0000-0002-8626-9822}{ORCID})

}
\keyword{internal}
