% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R libraries (One Line, Title Case)}
\description{
Provides functions for learning library and package information .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ma-ulloa/libminer}
  \item Report bugs at \url{https://github.com/ma-ulloa/libminer/issues}
}

}
\author{
\strong{Maintainer}: Maria Ulloa \email{alejaulloam@gmail.com}

Authors:
\itemize{
  \item Maria Ulloa \email{alejaulloam@gmail.com}
}

}
\keyword{internal}
