% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Okok, bla bla héhé je vais faire une phrase complète haha.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mathieu-desgagnes/libminer}
  \item Report bugs at \url{https://github.com/mathieu-desgagnes/libminer/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Desgagnés \email{mathieu.desgagnes@dfo-mpo.gc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
