% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanova-usgs/libminer}
  \item Report bugs at \url{https://github.com/mcanova-usgs/libminer/issues}
}

}
\author{
\strong{Maintainer}: Michael Canova \email{MichaelCanova@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
