% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R libraries}
\description{
Provides functions for learning about your R libs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/melimore86/libminer}
  \item Report bugs at \url{https://github.com/melimore86/libminer/issues}
}

}
\author{
\strong{Maintainer}: Mel M \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
