% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/miur1790/libminer}
  \item Report bugs at \url{https://github.com/miur1790/libminer/issues}
}

}
\author{
\strong{Maintainer}: Miren Urrutia \email{miren.urrutiaiturritza@su.se} (\href{https://orcid.org/0000-0002-3989-1803}{ORCID})

Authors:
\itemize{
  \item Miren Urrutia \email{miren.urrutiaiturritza@su.se} (\href{https://orcid.org/0000-0002-3989-1803}{ORCID})
}

}
\keyword{internal}
