% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjpinov/libminer}
  \item \url{https://mjpinov.github.io/libminer/}
  \item Report bugs at \url{https://github.com/mjpinov/libminer/issues}
}

}
\author{
\strong{Maintainer}: Maria Jose Pino \email{m.pinovsz@gmail.com}

}
\keyword{internal}
