% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickawbrown/libminer}
  \item \url{https://nickawbrown.github.io/libminer/}
  \item Report bugs at \url{https://github.com/nickawbrown/libminer/issues}
}

}
\author{
\strong{Maintainer}: Nick Brown \email{nickawbrown@gmail.com} (\href{https://orcid.org/0000-0002-9093-4098}{ORCID})

}
\keyword{internal}
