% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nlgarlic/libminer}
  \item Report bugs at \url{https://github.com/nlgarlic/libminer/issues}
}

}
\author{
\strong{Maintainer}: Nikki Lemire-Garlic \email{nlemiregarlic@unr.edu} (\href{https://orcid.org/00}{ORCID})

}
\keyword{internal}
