% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R libraries}
\description{
Provides functions to check the R packages
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Norruas/libminer}
  \item Report bugs at \url{https://github.com/Norruas/libminer/issues}
}

}
\author{
\strong{Maintainer}: Maryam Sahi \email{maryam.sahi@gmail.com}

Authors:
\itemize{
  \item Maryam Sahi \email{maryam.sahi@gmail.com}
}

}
\keyword{internal}
