% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Summarize R Libraries}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pandeyravi15/libminer}
  \item Report bugs at \url{https://github.com/pandeyravi15/libminer/issues}
}

}
\author{
\strong{Maintainer}: Ravi Pandey \email{ravisp1587@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
