% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Functions for learning about your R libraries and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/plandahl/libminer}
  \item Report bugs at \url{https://github.com/plandahl/libminer/issues}
}

}
\author{
\strong{Maintainer}: Paul Landahl \email{paul@landahl.net}

}
\keyword{internal}
