% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: What the Package Does (One Line, Title Case)}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RBelley/libminer}
  \item \url{https://rbelley.github.io/libminer/}
  \item Report bugs at \url{https://github.com/RBelley/libminer/issues}
}

}
\author{
\strong{Maintainer}: Renald Belley \email{renald.belley@dfo-mpo.gc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
