% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ritika-giri/libminer}
  \item Report bugs at \url{https://github.com/ritika-giri/libminer/issues}
}

}
\author{
\strong{Maintainer}: Ritika Giri \email{ritika.giri3@gmail.com} (\href{https://orcid.org/0000-0001-8838-0818}{ORCID})

}
\keyword{internal}
