% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robynforrest/libminer}
  \item Report bugs at \url{https://github.com/robynforrest/libminer/issues}
}

}
\author{
\strong{Maintainer}: Robyn E. Forrest \email{robyn.forrest@dfo-mpo.gc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
