% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robynforrest/libminer}
  \item Report bugs at \url{https://github.com/robynforrest/libminer/issues}
}

}
\author{
\strong{Maintainer}: Robyn Forrest (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
