% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/royfrancis/libminer}
  \item Report bugs at \url{https://github.com/royfrancis/libminer/issues}
}

}
\author{
\strong{Maintainer}: Roy Francis \email{first.last@example.com}

Authors:
\itemize{
  \item Roy Francis \email{first.last@example.com}
}

}
\keyword{internal}
