% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SantanaPC/libminer}
  \item Report bugs at \url{https://github.com/SantanaPC/libminer/issues}
}

}
\author{
\strong{Maintainer}: Pamela Santana \email{pamela.santana@biol.lu.se} (\href{https://orcid.org/0000-0002-7517-6671}{ORCID})

}
\keyword{internal}
