% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning code your R libraries, and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shffer/libminer}
  \item Report bugs at \url{https://github.com/shffer/libminer/issues}
}

}
\author{
\strong{Maintainer}: Shane Fernandez \email{s.fernandez@ecu.edu.au}

}
\keyword{internal}
