% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sinmojito/libminer}
  \item \url{https://sinmojito.github.io/libminer/}
  \item Report bugs at \url{https://github.com/sinmojito/libminer/issues}
}

}
\author{
\strong{Maintainer}: Mohit Panwar \email{mohitsingh5123@gmail.com} (\href{https://orcid.org/0009-0002-8866-8170}{ORCID})

Authors:
\itemize{
  \item Mohit Panwar \email{mohitsingh5123@gmail.com} (\href{https://orcid.org/0009-0002-8866-8170}{ORCID})
}

}
\keyword{internal}
