% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about both your R libraries , and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sstevens2/libminer}
  \item Report bugs at \url{https://github.com/sstevens2/libminer/issues}
}

}
\author{
\strong{Maintainer}: Sarah Stevens \email{sarah.stevens@wisc.edu} (\href{https://orcid.org/0000-0002-7040-548X}{ORCID})

}
\keyword{internal}
