% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\author{
\strong{Maintainer}: Stella Axelsson \email{stella.axelsson@scilifelab.se} (\href{https://orcid.org/0009-0002-0986-7850}{ORCID})

Authors:
\itemize{
  \item Stella Axelsson \email{stella.axelsson@scilifelab.se} (\href{https://orcid.org/0009-0002-0986-7850}{ORCID})
}

}
\keyword{internal}
