% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Locally Installed R Libraries}
\description{
Provides functions to explore locally-installed R libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephhazlitt/libminer}
  \item Report bugs at \url{https://github.com/stephhazlitt/libminer/issues}
}

}
\author{
\strong{Maintainer}: Stephanie Hazlitt \email{stephhazlitt@gmail.com} (\href{https://orcid.org/0000-0002-3161-2304}{ORCID})

}
\keyword{internal}
