% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R libraries}
\description{
Allows you to explore which R libraries you have and the number of packages they have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/suarezalvarez/libminer}
  \item Report bugs at \url{https://github.com/suarezalvarez/libminer/issues}
}

}
\author{
\strong{Maintainer}: Martín Suárez Álvarez \email{martin.suarez.alvarez@ki.se}

Authors:
\itemize{
  \item Martín Suárez Álvarez \email{martin.suarez.alvarez@ki.se}
}

}
\keyword{internal}
