% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R libraries}
\description{
Provides functions for learning about your R library and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Syi824/libminer}
  \item Report bugs at \url{https://github.com/Syi824/libminer/issues}
}

}
\author{
\strong{Maintainer}: Shirley Yi \email{syi@cpg.org}

}
\keyword{internal}
