% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Provides fucntions for learning about your R libraries, and the packages they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SylviaHao/libminer}
  \item \url{https://sylviahao.github.io/libminer/}
  \item Report bugs at \url{https://github.com/SylviaHao/libminer/issues}
}

}
\author{
\strong{Maintainer}: Sylvia Aponte-Hao \email{sylviahao@gmail.com} (\href{https://orcid.org/0000-0002-8940-4516}{ORCID})

}
\keyword{internal}
