% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: ExploreYour R libraries}
\description{
Provide functions to learn about R libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/theb8/libminer}
  \item \url{https://theb8.github.io/libminer/}
  \item Report bugs at \url{https://github.com/theb8/libminer/issues}
}

}
\author{
\strong{Maintainer}: Alex Theberge \email{theb8@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
