% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Explore Your R Libraries}
\description{
Basic Exploration of Your R Packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wagnerlmichael/libminer}
  \item Report bugs at \url{https://github.com/wagnerlmichael/libminer/issues}
}

}
\author{
\strong{Maintainer}: Michael Wagner \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
