% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer-package.R
\docType{package}
\name{libminer-package}
\alias{libminer}
\alias{libminer-package}
\title{libminer: Check the packages number in the libraries}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yiquanlu.github.io/libminer/}
}

}
\author{
\strong{Maintainer}: Yiquan Lu \email{6yiquan@gmail.com}

}
\keyword{internal}
