% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{libsummary}
\alias{libsummary}
\title{R Library Summary}
\usage{
libsummary()
}
\value{
A \code{data.frame} containing the count of
packages in each of the user's libraries
}
\description{
Provide a brief summary of the package
libraries on the machine
}
\examples{
libsummary()
}
